pyproject.toml
[project]
dependencies = [
  "accelerate>=1.8.0",
  "datasets>=3.6.0",
  "click>=8.1.7",
  "fabric>=3.2.2",
  "fastapi>=0.115.13",
  "isort>=6.0.1",
  "latex2sympy2>=1.9.1",
  "matplotlib>=3.10.3",
  "pebble>=5.1.1",
  "pyrootutils>=1.0.4",
  "pytest>=8.4.1",
  "python-dotenv>=1.1.1",
  "regex>=2024.11.6",
  "rich>=14.0.0",
  "ruff>=0.12.0",
  "tensorboard>=2.20.0",
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "wandb>=0.21.0",
  "word2number>=1.1",
  "scipy>=1.16.1",
  "scikit-learn>=1.7.1",
  "matplotlib-venn>=1.1.2",
  "seaborn>=0.13.2",
]

description = "Add your description here"
name = "search-and-learn"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/sal", "src/sal_experiments"]

[tool.uv.sources]
search-and-learn = {workspace = true}
trl = {git = "https://github.com/huggingface/trl.git"}

[dependency-groups]
deploy = [
  "editables>=0.5",
  "flash-attn>=2.8.2",
  "flashinfer-python>=0.2.11",
  "hatchling>=1.27.0",
  "matplotlib>=3.10.3",
  "nvidia-ml-py>=12.575.51",
  "transformers<4.54.0",
  "vllm==0.9.2",
]

[tool.pytest.ini_options]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
